refactor(audit): apply PR review suggestions
