+++ content/src/main/java/pt/estga/content/entities/MarkOccurrence.java
@@ -3,7 +3,7 @@ package pt.estga.content.entities;
 import jakarta.persistence.*;
 import lombok.*;
 import pt.estga.file.entities.MediaFile;
-import pt.estga.shared.entities.AuditedEntity;
+import pt.estga.shared.audit.AuditedEntity;
 import pt.estga.shared.utils.DoubleListConverter;
 
 import java.time.Instant;
