+++ content/src/main/java/pt/estga/content/mappers/MarkMapper.java
@@ -1,8 +1,10 @@
 package pt.estga.content.mappers;
 
+import org.mapstruct.BeanMapping;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingTarget;
+import org.mapstruct.NullValuePropertyMappingStrategy;
 import pt.estga.content.dtos.MarkListDto;
 import pt.estga.content.dtos.MarkUpdateDto;
 import pt.estga.content.entities.Mark;
@@ -21,6 +23,7 @@ public interface MarkMapper {
     @Mapping(source = "coverId", target = "cover.id")
     Mark updateDtoToEntity(MarkUpdateDto markDto);
 
+    @BeanMapping(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
     @Mapping(source = "coverId", target = "cover.id")
     void updateEntityFromDto(MarkUpdateDto dto, @MappingTarget Mark entity);
