+++ content/src/main/java/pt/estga/content/mappers/MonumentMapper.java
@@ -1,8 +1,10 @@
 package pt.estga.content.mappers;
 
+import org.mapstruct.BeanMapping;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import org.mapstruct.MappingTarget;
+import org.mapstruct.NullValuePropertyMappingStrategy;
 import pt.estga.content.dtos.*;
 import pt.estga.content.entities.Monument;
 import pt.estga.file.mappers.MediaFileMapper;
@@ -23,6 +25,7 @@ public interface MonumentMapper {
 
     Monument toEntity(MonumentRequestDto dto);
 
+    @BeanMapping(nullValuePropertyMappingStrategy = NullValuePropertyMappingStrategy.IGNORE)
     void updateEntityFromDto(MonumentRequestDto dto, @MappingTarget Monument entity);
 
 }
