+++ shared/pom.xml
@@ -39,7 +39,6 @@
         <dependency>
             <groupId>org.springframework.data</groupId>
             <artifactId>spring-data-envers</artifactId>
-            <version>4.0.1</version>
         </dependency>
     </dependencies>
